Add styles to label and radio group, rename files
