--- components/radio-group/WeeklyPickButton.tsx
@@ -1,17 +1,0 @@
-import { Label } from './Label';
-import { RadioGroup, RadioGroupItem } from './RadioGroup';
-
-export const WeeklyPickButton = () => {
-  return (
-    <RadioGroup defaultValue="option-one">
-      <div className="flex items-center space-x-2">
-        <RadioGroupItem value="option-one" id="option-one" />
-        <Label htmlFor="option-one">Option One</Label>
-      </div>
-      <div className="flex items-center space-x-2">
-        <RadioGroupItem value="option-two" id="option-two" />
-        <Label htmlFor="option-two">Option Two</Label>
-      </div>
-    </RadioGroup>
-  );
-};
