+++ components/weekly-pick-button/Label.tsx
@@ -7,7 +7,7 @@ import { cva, type VariantProps } from 'class-variance-authority';
 import { cn } from '@/lib/utils';
 
 const labelVariants = cva(
-  'text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70',
+  'text-lg flex cursor-pointer gap-2 items-center rounded-xl border-2 border-zinc-800 font-medium leading-none peer-aria-checked:border-orange-600 py-4 px-3 w-full peer-hover:bg-zinc-800 peer-hover:text-zinc-50 transition',
 );
 
 const Label = React.forwardRef<
